refactor(serial): fold framer.rs into framing.rs
`framer.rs` next to `framing.rs` was a confusable pair for one subject: the
COBS codec and that codec behind core's `Framer` trait. They are now one
module, with the module doc naming the two layers and why only the second is
feature-gated.

The gate moves from the module declaration onto the items that need it —
`CobsFramer`, the chunk sizes and the `FramedConnection` aliases name
`aimdb_core::session`, which core gates on `connector-session`. `encode_frame`
and `FrameAccumulator` never needed it and stay ungated, so the codec still
builds with no runtime feature at all.

Public paths change from `framer::*` to `framing::*`; the items keep their
names.

Verified: clippy `-D warnings` for tokio, the `thumbv7em` embassy
cross-compile (which type-checks `_same_framed_connection_serves_the_uart`,
the assertion that the two runtime paths have not diverged), and no runtime
feature at all; serial tests; rustdoc on both runtimes; `make doc`;
`cargo fmt --check`.
